requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Authentication (custom implementation using hashlib, no external libraries needed)

# Validation
email-validator>=2.0.0

# Request handling
python-multipart>=0.0.6
aiofiles>=23.2.0

# PDF parsing (for scripts)
pdfplumber>=0.10.0

# Environment variables
python-dotenv>=1.0.0
